better：心流升级，大大减少了复读情况，并且灵活调用工具来实现知识和记忆检索
